+++ shared/src/lib.rs
@@ -1,14 +1,15 @@
+pub mod amm_resource;
 pub mod arguments;
 pub mod conversions;
 pub mod current_block;
 pub mod gas_price_estimation;
 pub mod metrics;
+pub mod pool_fetching;
 pub mod price_estimate;
 pub mod time;
 pub mod token_info;
 pub mod tracing;
 pub mod transport;
-pub mod uniswap_pool;
 pub mod uniswap_solver;
 
 pub type Web3 = web3::Web3<transport::LoggingTransport<web3::transports::Http>>;
